Fixed Timing Errors
Occasionally, the extension would fail to retrieve the time until the next boss. To address this, I've implemented a function that continuously verifies whether the loop responsible for fetching this information from my server is operational. If the loop is not running, the function will automatically restart the extension's processes and obtain the time once more.